+++ components/mdx-remote/mdx-remote.tsx
@@ -1,3 +1,11 @@
 "use client";
 
-export { MDXRemote } from "next-mdx-remote";
+import { components } from "components/mdx";
+
+import { MDXRemote as OriginalMDXRemote } from "next-mdx-remote";
+
+export const MDXRemote = ({ source, ...props }: any) => {
+  return (
+    <OriginalMDXRemote {...props} components={components} source={source} />
+  );
+};
